+++ src/main/java/com/unerp/service/product/ProductUpdateService.java
@@ -39,7 +39,7 @@ public class ProductUpdateService {
         String existingName = existingProduct.getName();
         String existingDescription = existingProduct.getDescription();
         Integer existingStock = existingProduct.getStock();
-        Double existingPrice = existingProduct.getPrice();
+        Double existingPrice = existingProduct.getPrice().doubleValue();
         String existingBatch = existingProduct.getBatch();
         LocalDate existingExpirationDate = existingProduct.getExpirationDate();
         Integer existingSupplierId = existingProduct.getSupplierId();
